+++ src/components/Filter.tsx
@@ -50,6 +50,10 @@ export default class Filter extends React.Component<FilterProps, FilterState> {
     this.props.flipFilter();
   }
 
+  /*
+    This function sets the state of the Filter component to open or close, 
+    which in turn tells the Panel component whether to render its content or not.
+   */
   handleFilterPanelClick() {
     this.setState((prevState) => ({
       ...prevState,